vold_prepare_subdirs: Take over /data/system_ce/#/appsearch mkdir
In order to ensure that the appropriate SELinux label is created
for AppSearch, vold_prepare_subdirs needs to handle the top level
directory creation instead of allowing AppSearch to do it itself.
Without this change, when AppSearch attempted to create the
appsearch directory, it would be assigned
u:object_r:system_data_file:s0, which is the default label for
files on the /data partition.

Flag: EXEMPT
Test: atest AppSearchSessionCtsTest
Test: atest --rebuild-module-info CtsAppSearchTestCases AppSearchServicesTests AppSearchMockingServicesTests AppSearchCoreTests
Test: atest --rebuild-module-info CtsAppSearchHostTestCases
Change-Id: I828b772884f102e2d73bdc18f94fce6767ea532c