Reapply "Delete the checkin directory with the wrong context"
This is a reapplication of commit b2ee9e0771a350d0c6331493c313fe37435facb8,
which was reverted in dd789a0e61a25b9bb5e889b70b9285ddb6f5382f.

The original commit message is as follows:

  Delete the checkin directory with the wrong context

  http://aosp/1845900 creates the directory with the wrong permission
  and context. And when we attempt to fix it in http://aosp/1860276, the
  device would fail to boot if the device is already on the bad build.

  As a temporarily fix, already delete that checkin directory in vold. And
  we can revert the deletion when the droidfood daily polulation gets out
  of the bad state.

  Bug: 203742483
  Test: Update from TP1A.211016.001 and make sure the boot doesn't fail

For unknown reasons, this directory deletion still seems necessary.
According to b/461722083#comment7 , the restorecon operation is failing
when changing the label of the "checkin" directory from
u:object_r:checkin_data_file:s0 to u:object_r:checkin_data_file:s0:c512,c768.
It's possible that a different restorecon is stripping the MLS label
from the directory, and the attempt to reapply the label is failing.

The safest thing to do is to revert the revert until we have a better
understanding of this bug.

Bug: 461722083
Test: Compiles and boots
Flag: EXEMPT bugfix
Change-Id: I39aadd2a08bb4222b82cba56ab6dc2cd070444c3